changes due to change requests
* reverted translation field names ("BtnLabel..." to "Label...")
* separate files for the tools
* fix text formating in crosshair tool
* some smaller non-functional refactoring
