+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/Messages.java
@@ -591,7 +591,7 @@ public class Messages extends NLS
     public static String LabelConvertBuySellIntoDeliveryTransactions;
     public static String LabelCopyToClipboard;
     public static String LabelCountry;
-    public static String BtnLabelCrosshair;
+    public static String LabelCrosshair;
     public static String LabelCurrencies;
     public static String LabelCurrencyConverter;
     public static String LabelCurrentDate;
@@ -787,7 +787,7 @@ public class Messages extends NLS
     public static String LabelKeyIndicators;
     public static String LabelMaxDrawdown;
     public static String LabelMaxDrawdownDuration;
-    public static String BtnLabelMeasureDistance;
+    public static String LabelMeasureDistance;
     public static String LabelMetricCheckInterval;
     public static String LabelMetric;
     public static String LabelMetricDays;
