+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages.properties
@@ -1192,7 +1192,7 @@ LabelCopyToClipboard = Copy to Clipboard
 
 LabelCountry = Country
 
-BtnLabelCrosshair=Crosshair
+LabelCrosshair=Crosshair
 
 LabelCurrencies = Currencies
 
@@ -1382,7 +1382,7 @@ LabelMaxDrawdown = Maximum Drawdown
 
 LabelMaxDrawdownDuration = Max Drawdown Duration
 
-BtnLabelMeasureDistance = Measure distance
+LabelMeasureDistance = Measure distance
 
 LabelMetric = Metric
