+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_cs.properties
@@ -1186,7 +1186,7 @@ LabelCopyToClipboard = Zkop\u00EDrovat do schr\u00E1nky
 
 LabelCountry = Zem\u011B
 
-BtnLabelCrosshair = Crosshair
+LabelCrosshair = Crosshair
 
 LabelCurrencies = M\u011Bny
 
@@ -1376,7 +1376,7 @@ LabelMaxDrawdown = Maxim\u00E1ln\u00ED \u010Derp\u00E1n\u00ED
 
 LabelMaxDrawdownDuration = Maxim\u00E1ln\u00ED doba \u010Derp\u00E1n\u00ED
 
-BtnLabelMeasureDistance = M\u011B\u0159en\u00ED vzd\u00E1lenosti
+LabelMeasureDistance = M\u011B\u0159en\u00ED vzd\u00E1lenosti
 
 LabelMetric = Metrika
