+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_da.properties
@@ -1185,7 +1185,7 @@ LabelCopyToClipboard = Kopier til udklipsholder
 
 LabelCountry = Land
 
-BtnLabelCrosshair = Tr\u00E5dkors
+LabelCrosshair = Tr\u00E5dkors
 
 LabelCurrencies = Valutaer
 
@@ -1375,7 +1375,7 @@ LabelMaxDrawdown = Maksimale Tilbagefald
 
 LabelMaxDrawdownDuration = Maksimale Tilbagefald Varighed
 
-BtnLabelMeasureDistance = M\u00E5ling af afstand
+LabelMeasureDistance = M\u00E5ling af afstand
 
 LabelMetric = Metrisk
