+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_de.properties
@@ -1185,7 +1185,7 @@ LabelCopyToClipboard = In Zwischenablage kopieren
 
 LabelCountry = Land
 
-BtnLabelCrosshair = Fadenkreuz
+LabelCrosshair = Fadenkreuz
 
 LabelCurrencies = W\u00E4hrungen
 
@@ -1375,7 +1375,7 @@ LabelMaxDrawdown = Maximaler Drawdown
 
 LabelMaxDrawdownDuration = Maximale Drawdown Duration
 
-BtnLabelMeasureDistance = Abstand messen
+LabelMeasureDistance = Abstand messen
 
 LabelMetric = Metrik
