+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_es.properties
@@ -1185,7 +1185,7 @@ LabelCopyToClipboard = Copiar al portapapeles
 
 LabelCountry = Pa\u00EDs
 
-BtnLabelCrosshair = Crosshair
+LabelCrosshair = Crosshair
 
 LabelCurrencies = Monedas
 
@@ -1375,7 +1375,7 @@ LabelMaxDrawdown = Drawdown m\u00E1ximo
 
 LabelMaxDrawdownDuration = M\u00E1xima duraci\u00F3n de Drawdown
 
-BtnLabelMeasureDistance = Modo de Medici\u00F3n
+LabelMeasureDistance = Modo de Medici\u00F3n
 
 LabelMetric = M\u00E9trico
