+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_fr.properties
@@ -1186,7 +1186,7 @@ LabelCopyToClipboard = Copier dans le presse-papier
 
 LabelCountry = Pays
 
-BtnLabelCrosshair = R\u00E9ticule
+LabelCrosshair = R\u00E9ticule
 
 LabelCurrencies = Devises
 
@@ -1370,7 +1370,7 @@ LabelMaxDrawdown = Maximum Drawdown
 
 LabelMaxDrawdownDuration = Dur\u00E9e Max. Drawdown
 
-BtnLabelMeasureDistance = Mesurer la distance
+LabelMeasureDistance = Mesurer la distance
 
 LabelMetric = M\u00E9trique
