+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_it.properties
@@ -1185,7 +1185,7 @@ LabelCopyToClipboard = Copia negli appunti
 
 LabelCountry = Nazione
 
-BtnLabelCrosshair = Mirino
+LabelCrosshair = Mirino
 
 LabelCurrencies = Valute
 
@@ -1375,7 +1375,7 @@ LabelMaxDrawdown = Massimo Drawdown
 
 LabelMaxDrawdownDuration = Durata Max Drawdown
 
-BtnLabelMeasureDistance = Misura distanza
+LabelMeasureDistance = Misura distanza
 
 LabelMetric = Metrico
