+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_nl.properties
@@ -1185,7 +1185,7 @@ LabelCopyToClipboard = Kopi\u00EBren naar klembord
 
 LabelCountry = Land
 
-BtnLabelCrosshair = Dradenkruis
+LabelCrosshair = Dradenkruis
 
 LabelCurrencies = Valuta
 
@@ -1375,7 +1375,7 @@ LabelMaxDrawdown = Maximale daling
 
 LabelMaxDrawdownDuration = Max. duur van de daling
 
-BtnLabelMeasureDistance = Afstand meten
+LabelMeasureDistance = Afstand meten
 
 LabelMetric = Metriek
