+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_pl.properties
@@ -1185,7 +1185,7 @@ LabelCopyToClipboard = Skopiuj do schowka
 
 LabelCountry = Kraj
 
-BtnLabelCrosshair = Celownik
+LabelCrosshair = Celownik
 
 LabelCurrencies = Waluty
 
@@ -1375,7 +1375,7 @@ LabelMaxDrawdown = Maksymalna wyp\u0142ata
 
 LabelMaxDrawdownDuration = Maksymalny czas wyp\u0142aty
 
-BtnLabelMeasureDistance = Zmierz odleg\u0142o\u015B\u0107
+LabelMeasureDistance = Zmierz odleg\u0142o\u015B\u0107
 
 LabelMetric = Metryczny
