+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_pt.properties
@@ -1185,7 +1185,7 @@ LabelCopyToClipboard = Copiar para a \u00E1rea de transfer\u00EAncia
 
 LabelCountry = Pa\u00EDs
 
-BtnLabelCrosshair = Mira
+LabelCrosshair = Mira
 
 LabelCurrencies = Moedas
 
@@ -1375,7 +1375,7 @@ LabelMaxDrawdown = Drawdown m\u00E1ximo
 
 LabelMaxDrawdownDuration = Dura\u00E7\u00E3o m\u00E1xima do drawdown
 
-BtnLabelMeasureDistance = Medir dist\u00E2ncia
+LabelMeasureDistance = Medir dist\u00E2ncia
 
 LabelMetric = M\u00E9trica
