+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_ru.properties
@@ -1185,7 +1185,7 @@ LabelCopyToClipboard = \u041A\u043E\u043F\u0438\u0440\u043E\u0432\u0430\u0442\u0
 
 LabelCountry = \u0421\u0442\u0440\u0430\u043D\u0430
 
-BtnLabelCrosshair = \u041F\u0435\u0440\u0435\u043A\u0440\u0435\u0441\u0442\u0438\u0435
+LabelCrosshair = \u041F\u0435\u0440\u0435\u043A\u0440\u0435\u0441\u0442\u0438\u0435
 
 LabelCurrencies = \u0412\u0430\u043B\u044E\u0442\u044B
 
@@ -1373,7 +1373,7 @@ LabelMaxDrawdown = \u041C\u0430\u043A\u0441\u0438\u043C\u0430\u043B\u044C\u043D\
 
 LabelMaxDrawdownDuration = \u041C\u0430\u043A\u0441\u0438\u043C\u0430\u043B\u044C\u043D\u0430\u044F \u0434\u043B\u0438\u0442\u0435\u043B\u044C\u043D\u043E\u0441\u0442\u044C \u043F\u0440\u043E\u0441\u0430\u0434\u043A\u0438
 
-BtnLabelMeasureDistance = \u0418\u0437\u043C\u0435\u0440\u044C\u0442\u0435 \u0440\u0430\u0441\u0441\u0442\u043E\u044F\u043D\u0438\u0435
+LabelMeasureDistance = \u0418\u0437\u043C\u0435\u0440\u044C\u0442\u0435 \u0440\u0430\u0441\u0441\u0442\u043E\u044F\u043D\u0438\u0435
 
 LabelMetric = \u041C\u0435\u0442\u0440\u0438\u043A\u0430
