+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_sk.properties
@@ -1185,7 +1185,7 @@ LabelCopyToClipboard = Kop\u00EDrovanie do schr\u00E1nky
 
 LabelCountry = Krajina
 
-BtnLabelCrosshair = Kri\u017Eovatka
+LabelCrosshair = Kri\u017Eovatka
 
 LabelCurrencies = Meny
 
@@ -1375,7 +1375,7 @@ LabelMaxDrawdown = Maxim\u00E1lne \u010Derpanie
 
 LabelMaxDrawdownDuration = Maxim\u00E1lna d\u013A\u017Eka \u010Derpania
 
-BtnLabelMeasureDistance = Meranie vzdialenosti
+LabelMeasureDistance = Meranie vzdialenosti
 
 LabelMetric = Metrick\u00E9
