+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_zh.properties
@@ -1185,7 +1185,7 @@ LabelCopyToClipboard = \u590D\u5236\u81F3\u526A\u8D34\u677F
 
 LabelCountry = \u56FD\u5BB6\u6216\u5730\u533A
 
-BtnLabelCrosshair = \u5341\u5B57\u51C6\u7EBF
+LabelCrosshair = \u5341\u5B57\u51C6\u7EBF
 
 LabelCurrencies = \u8D27\u5E01
 
@@ -1375,7 +1375,7 @@ LabelMaxDrawdown = \u6700\u5927\u56DE\u64A4
 
 LabelMaxDrawdownDuration = \u6700\u957F\u56DE\u64A4\u671F
 
-BtnLabelMeasureDistance = \u6D4B\u91CF\u8DDD\u79BB
+LabelMeasureDistance = \u6D4B\u91CF\u8DDD\u79BB
 
 LabelMetric = \u5EA6\u91CF
